requirements.txt
Django==5.0.1
channels==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0